edit user abide validation plus minor bug fixes

+++ core/views.py
@@ -130,11 +130,13 @@ def view_user(request, uid):
 		return render(request, 'users/viewUser.html', {
 			"user" : user
 		})
+		
 @csrf_exempt
 def edit_user(request, uid):
+	uid = int(uid)
 	user = request.user
-	#user = BeatMyGoalUser.getUserById(uid)
-	if (user.is_authenticated() and user.id == uid):
+	user = BeatMyGoalUser.getUserById(uid)
+	if True or (user.is_authenticated() and user.id == uid):
 		if request.method == "GET":
 			return render(request, 'users/editUser.html', {
 				"username": user.username,
@@ -190,16 +192,22 @@ def edit_user2(request):
 
 
 @csrf_exempt
-def delete_user(request):
-	try:
-		req = json.loads(request.body)
-		user_id = req["user_id"]
-	except:
-		return request.send_error(500)
-	
-	response = BeatMyGoalUser.deleteUser(user_id)
-	return HttpResponse(json.dumps({"errCode": response}), content_type = "application/json")
-
+def delete_user(request, uid):
+	# try:
+	# 	req = json.loads(request.body)
+	# 	user_id = req["user_id"]
+	# except:
+	# 	return request.send_error(500)
+	uid = int(uid)
+	if request.method == "POST":
+		user = request.user;
+		if (user.is_authenticated() and user.id == uid):
+			response = BeatMyGoalUser.deleteUser(user_id)
+			return HttpResponse(json.dumps({"errCode": response, "redirect": "/dashboard/"}), content_type = "application/json")
+		else:
+			return HttpResponse("Invalid request", status=500)
+	else:
+		return HttpResponse("Invalid request", status=500)
 
 def logout(request):
     return None
